Use one of scope handler
