+++ packages/cursorless-engine/src/processTargets/modifiers/scopeHandlers/NotebookCellScopeHandler.ts
@@ -1,143 +1,70 @@
 import {
-  Range,
   type Direction,
-  type NotebookCell,
   type Position,
   type ScopeType,
   type TextEditor,
 } from "@cursorless/common";
 import type { LanguageDefinitions } from "../../../languages/LanguageDefinitions";
-import { ide } from "../../../singletons/ide.singleton";
-import { NotebookCellTarget } from "../../targets";
 import { BaseScopeHandler } from "./BaseScopeHandler";
+import { NotebookCellApiScopeHandler } from "./NotebookCellApiScopeHandler";
+import { OneOfScopeHandler } from "./OneOfScopeHandler";
 import type { TargetScope } from "./scope.types";
 import type {
   ComplexScopeType,
   ScopeHandler,
   ScopeIteratorRequirements,
 } from "./scopeHandler.types";
+import type { ScopeHandlerFactory } from "./ScopeHandlerFactory";
 
 export class NotebookCellScopeHandler extends BaseScopeHandler {
   public readonly scopeType = { type: "notebookCell" } as const;
   protected isHierarchical = false;
-  private readonly languageScopeHandler: ScopeHandler | undefined;
+  private readonly scopeHandler: ScopeHandler;
+
+  get iterationScopeType(): ScopeType | ComplexScopeType {
+    return this.scopeHandler.iterationScopeType;
+  }
 
   constructor(
+    scopeHandlerFactory: ScopeHandlerFactory,
     languageDefinitions: LanguageDefinitions,
     _scopeType: ScopeType,
-    private languageId: string,
+    languageId: string,
   ) {
     super();
 
-    this.languageScopeHandler = languageDefinitions
-      .get(this.languageId)
-      ?.getScopeHandler(this.scopeType);
+    this.scopeHandler = (() => {
+      const apiScopeHandler = new NotebookCellApiScopeHandler();
+
+      const languageScopeHandler = languageDefinitions
+        .get(languageId)
+        ?.getScopeHandler(this.scopeType);
+
+      if (languageScopeHandler == null) {
+        return apiScopeHandler;
+      }
+
+      return OneOfScopeHandler.createFromScopeHandlers(
+        scopeHandlerFactory,
+        {
+          type: "oneOf",
+          scopeTypes: [
+            languageScopeHandler.scopeType,
+            apiScopeHandler.scopeType,
+          ],
+        },
+        [languageScopeHandler, apiScopeHandler],
+        languageId,
+      );
+    })();
   }
 
-  get iterationScopeType(): ScopeType | ComplexScopeType {
-    if (this.languageScopeHandler != null) {
-      return this.languageScopeHandler.iterationScopeType;
-    }
-    return { type: "document" };
-  }
-
-  *generateScopeCandidates(
+  generateScopeCandidates(
     editor: TextEditor,
     position: Position,
     direction: Direction,
     hints: ScopeIteratorRequirements,
   ): Iterable<TargetScope> {
-    if (this.languageScopeHandler != null) {
-      yield* this.languageScopeHandler.generateScopes(
-        editor,
-        position,
-        direction,
-        hints,
-      );
-    }
-
-    const cells = getNotebookCells(editor, position, direction, hints);
-
-    for (const cell of cells) {
-      yield createTargetScope(cell);
-    }
+    return this.scopeHandler.generateScopes(editor, position, direction, hints);
   }
-}
-
-function getNotebookCells(
-  editor: TextEditor,
-  position: Position,
-  direction: Direction,
-  hints: ScopeIteratorRequirements,
-) {
-  const nb = getNotebook(editor);
-
-  if (nb == null) {
-    return [];
-  }
-
-  const { notebook, cell } = nb;
-
-  if (hints.containment === "required") {
-    return [cell];
-  }
-
-  if (
-    hints.containment === "disallowed" ||
-    hints.containment === "disallowedIfStrict"
-  ) {
-    return direction === "forward"
-      ? notebook.cells.slice(cell.index + 1)
-      : notebook.cells.slice(0, cell.index).reverse();
-  }
-
-  // Every scope
-  if (hints.distalPosition != null) {
-    const searchRange = new Range(position, hints.distalPosition);
-    if (searchRange.isRangeEqual(editor.document.range)) {
-      return notebook.cells;
-    }
-  }
-
-  return direction === "forward"
-    ? notebook.cells.slice(cell.index)
-    : notebook.cells.slice(0, cell.index + 1).reverse();
-}
-
-function getNotebook(editor: TextEditor) {
-  const uri = editor.document.uri.toString();
-  for (const notebook of ide().visibleNotebookEditors) {
-    for (const cell of notebook.cells) {
-      if (cell.document.uri.toString() === uri) {
-        return { notebook, cell };
-      }
-    }
-  }
-  return undefined;
-}
-
-function createTargetScope(cell: NotebookCell): TargetScope {
-  const editor = getEditor(cell);
-  const contentRange = editor.document.range;
-  return {
-    editor,
-    domain: contentRange,
-    getTargets: (isReversed: boolean) => [
-      new NotebookCellTarget({
-        editor,
-        isReversed,
-        contentRange,
-      }),
-    ],
-  };
-}
-
-function getEditor(cell: NotebookCell) {
-  const uri = cell.document.uri.toString();
-  for (const editor of ide().visibleTextEditors) {
-    if (editor.document.uri.toString() === uri) {
-      return editor;
-    }
-  }
-  throw new Error("Editor not found notebook cell");
 }
